+++ src/bittytax/conv/parsers/__init__.py
@@ -53,6 +53,7 @@ from . import (
     koinly,
     kraken,
     kucoin,
+    lbank,
     ledgerlive,
     liquid,
     mercatox,